v1/Gateway/pyproject.toml
[project]
name = "persons"
version = "1.0.1"
description = "..."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "typing-extensions==4.12.2",
    "sqlalchemy==2.0.41",
    "sqlalchemy-jsonfield==1.0.2",
    "sqlalchemy-utils==0.41.2",
    "python-dotenv==1.0.1",
    "fastapi==0.115.4",
    "fastapi-cli==0.0.5",
    "asyncpg==0.29.0",
    "greenlet>=3.0.0",
    "redis>=5.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
egg-base = ".cache/egg-info"

[tool.setuptools.packages.find]
include=["app*"]
exclude=[
    "uplatform",
    "ci"
]

[tool.uv]
default-groups = "all"

[dependency-groups]
dev = [
    "mypy>=1.18.1",
    "openpyxl>=3.1.5",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.13.0",
    "httpx>=0.27.0",
    "pytest-cov>=6.0.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"
include = [
    "pyproject.toml",
    "app/**/*.py",
    "tests/**/*.py"
]
exclude = [
    "uplatform",
    "ci",
    ".vscode",
    ".venv",
    "*.pytest_cache*.",
    ".cursor",
    ".docs",
    "__pycache__"
]

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors / https://docs.astral.sh/ruff/rules/#error-e
    "W", # pycodestyle warnings / https://docs.astral.sh/ruff/rules/#warning-w
    "F", # pyflakes / https://docs.astral.sh/ruff/rules/#pyflakes-f
    "I", # isort / https://docs.astral.sh/ruff/rules/#isort-i
    "UP", # pyupgrade / https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "D", # pydocstyle / https://docs.astral.sh/ruff/settings/#lintpydocstyle
    "N801", # Class name {name} should use CapWords convention
    "N802", # Function name {name} should be lowercase
    "N803", # Argument name {name} should be lowercase
    "N804", # First argument of a class method should be named cls
    "N805", # First argument of a method should be named self
    "S", # flake8-bandit
    "C4", # flake8-comprehensions для лучших list/dict comprehensions
    "SIM", # flake8-simplify для упрощения кода
]
ignore = [
    "F405",  # pyflakes: name may be undefined due to star import
    "F403",  # pyflakes: star import used
    "D417",  # pydocstyle: missing argument descriptions in the docstring
    "S501",  # bandit: request with no timeout
    "S608",  # bandit: possible SQL injection
    "S101",  # bandit: use of assert detected
    "D202",  # pydocstyle: no blank lines allowed after function docstring
    "D100",  # pydocstyle: missing docstring in public module
    "D400",  # pydocstyle: first line should end with a period
    "D205",  # pydocstyle: 1 blank line required between summary line and description
    "D104",   # pydocstyle: missing docstring in public package
    "D103",
    "D101",
    "D102",
    "D106",
    "D107"
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = "dynamic"
skip-magic-trailing-comma = false
preview = true  # Включает экспериментальные функции форматирования
line-ending = "lf"  # Использует LF для консистентности

[tool.ruff.lint.isort]
lines-between-types = 1
length-sort=true
combine-as-imports = true
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]
known-first-party = ["app", "alembic"]
known-third-party = [
"opentelemetry",
"dotenv",
"fastapi",
"sqlalchemy",
"pandas",
"numpy",
"requests",
"httpx",
"asyncpg",
"regex",
"yql",
"clickhouse_connect",
"yandex_yt",
"tvmauth",
"pydantic",
"procrastinate",
"psycopg",
]

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.ruff.lint.flake8-bandit]
check-typed-exception = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "asyncio: marks tests as async (deselect with '-m \"not asyncio\"')",
]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
namespace_packages = true
explicit_package_bases = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_decorators = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
disable_error_code = ["misc", "import-untyped"]
exclude = [
    ".*/test_.*\\.py"
]
